requirements.txt
click==7.1.2
Flask==1.1.2
Flask-DB2==0.0.10
ibm-db==3.0.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
python-dotenv==0.15.0
Werkzeug==1.0.1
